getLevelIcon moved to the utils folder for better modularity and tests for this function have been added.
