+++ src/components/technical-test/TechnicalTestCard.tsx
@@ -2,27 +2,12 @@ import { TechnicalTest } from "../../types/TechnicalTest";
 import XnixCalendar from "../../assets/xnix-calendar.svg";
 import { asideContentForTechnicalTest } from "../Layout/aside/asideContent";
 import { Link } from "react-router";
-import testLevel1 from "../../assets/testsLevel.svg";
-import testLevel2 from "../../assets/testsLevel2.svg";
-import testLevel3 from "../../assets/testsLevel3.svg";
 import { Clock, Heart } from "lucide-react";
+import { getLevelIcon } from "../../utils/getLevelIcon";
 
 interface TechnicalTestCardProps {
   test: TechnicalTest;
 }
-
-const getLevelIcon = (title: string) => {
-  const hash = Array.from(title).reduce(
-    (sum, char) => sum + char.charCodeAt(0),
-    0,
-  );
-
-  const level = (hash % 3) + 1;
-
-  if (level === 1) return testLevel1;
-  if (level === 2) return testLevel2;
-  return testLevel3;
-};
 
 const TechnicalTestCard = ({ test }: TechnicalTestCardProps) => {
   const language = asideContentForTechnicalTest.find(
